+++ test/draft3_test.rb
@@ -3,7 +3,7 @@ require File.expand_path('../support/test_helper', __FILE__)
 
 class Draft3Test < Minitest::Test
   def validation_errors(schema, data, options)
-    super(schema, data, :version => :draft3)
+    super(schema, data, options.merge(:version => :draft3))
   end
 
   def exclusive_minimum
@@ -31,7 +31,9 @@ class Draft3Test < Minitest::Test
   include ObjectValidation::AdditionalPropertiesTests
   include ObjectValidation::PatternPropertiesTests
 
-  include StrictValidation
+  include StrictValidation::PropertiesTests
+  include StrictValidation::AdditionalPropertiesTests
+  include StrictValidation::PatternPropertiesTests
 
   include StringValidation::ValueTests
   include StringValidation::FormatTests
@@ -66,8 +68,7 @@ class Draft3Test < Minitest::Test
     }
 
     data["a"] = [{"b" => 5}, {"c" => "foo"}]
-    errors = JSON::Validator.fully_validate(schema, data)
-    assert(errors.empty?, errors.join("\n"))
+    assert_valid schema, data
 
     # This should actually pass, because this matches the first schema in the union
     data["a"] << {"c" => false}
@@ -109,13 +110,13 @@ class Draft3Test < Minitest::Test
     }
 
     data = {"a" => "a"}
-    assert(JSON::Validator.validate(schema,data,:strict => true))
+    assert_valid schema,data,:strict => true
 
     data = {"b" => "b"}
-    assert(!JSON::Validator.validate(schema,data,:strict => true))
+    refute_valid schema,data,:strict => true
 
     data = {"a" => "a", "b" => "b"}
-    assert(JSON::Validator.validate(schema,data,:strict => true))
+    assert_valid schema,data,:strict => true
   end
 
   def test_strict_properties_additional_props
@@ -129,19 +130,19 @@ class Draft3Test < Minitest::Test
     }
 
     data = {"a" => "a"}
-    assert(!JSON::Validator.validate(schema,data,:strict => true))
+    refute_valid schema,data,:strict => true
 
     data = {"b" => "b"}
-    assert(!JSON::Validator.validate(schema,data,:strict => true))
+    refute_valid schema,data,:strict => true
 
     data = {"a" => "a", "b" => "b"}
-    assert(JSON::Validator.validate(schema,data,:strict => true))
+    assert_valid schema,data,:strict => true
 
     data = {"a" => "a", "b" => "b", "c" => "c"}
-    assert(!JSON::Validator.validate(schema,data,:strict => true))
+    refute_valid schema,data,:strict => true
 
     data = {"a" => "a", "b" => "b", "c" => 3}
-    assert(JSON::Validator.validate(schema,data,:strict => true))
+    assert_valid schema,data,:strict => true
   end
 
   def test_strict_properties_pattern_props
@@ -155,25 +156,25 @@ class Draft3Test < Minitest::Test
     }
 
     data = {"a" => "a"}
-    assert(!JSON::Validator.validate(schema,data,:strict => true))
+    refute_valid schema,data,:strict => true
 
     data = {"b" => "b"}
-    assert(!JSON::Validator.validate(schema,data,:strict => true))
+    refute_valid schema,data,:strict => true
 
     data = {"a" => "a", "b" => "b"}
-    assert(JSON::Validator.validate(schema,data,:strict => true))
+    assert_valid schema,data,:strict => true
 
     data = {"a" => "a", "b" => "b", "c" => "c"}
-    assert(!JSON::Validator.validate(schema,data,:strict => true))
+    refute_valid schema,data,:strict => true
 
     data = {"a" => "a", "b" => "b", "c" => 3}
-    assert(!JSON::Validator.validate(schema,data,:strict => true))
+    refute_valid schema,data,:strict => true
 
     data = {"a" => "a", "b" => "b", "23 taco" => 3}
-    assert(JSON::Validator.validate(schema,data,:strict => true))
+    assert_valid schema,data,:strict => true
 
     data = {"a" => "a", "b" => "b", "23 taco" => "cheese"}
-    assert(!JSON::Validator.validate(schema,data,:strict => true))
+    refute_valid schema,data,:strict => true
   end
 
   def test_disallow
@@ -231,7 +232,7 @@ class Draft3Test < Minitest::Test
     }
 
     assert_valid schema, data
-    assert(!JSON::Validator.validate(schema2,data))
+    refute_valid schema2,data
 
     schema["extends"] = schema2
 
@@ -246,14 +247,14 @@ class Draft3Test < Minitest::Test
     }
 
     data = [{"a" => 1},{"a" => 2},{"a" => 3}]
-    assert(JSON::Validator.validate(schema,data,:list => true))
+    assert_valid schema,data,:list => true
     refute_valid schema, data
 
     data = {"a" => 1}
-    assert(!JSON::Validator.validate(schema,data,:list => true))
+    refute_valid schema,data,:list => true
 
     data = [{"a" => 1},{"b" => 2},{"a" => 3}]
-    assert(!JSON::Validator.validate(schema,data,:list => true))
+    refute_valid schema,data,:list => true
   end
 
 
@@ -318,9 +319,9 @@ class Draft3Test < Minitest::Test
         "properties" => { "a" => {"type" => "string", "format" => "uri"}}
     }
 
-    assert(JSON::Validator.validate(schema,data1))
-    assert(!JSON::Validator.validate(schema,data2))
-    assert(JSON::Validator.validate(schema,data3))
+    assert_valid schema,data1
+    refute_valid schema,data2
+    assert_valid schema,data3
   end
 
 
@@ -332,7 +333,7 @@ class Draft3Test < Minitest::Test
     }
 
     data = {"a" => "taco"}
-    assert(!JSON::Validator.validate(schema, data))
+    refute_valid schema, data
 
     schema = {
       "$schema" => "http://json-schema.org/draft-03/schema#",
@@ -391,7 +392,7 @@ class Draft3Test < Minitest::Test
     data = {:b => 2}
     assert_valid schema, data
     assert_nil(data["a"])
-    assert(JSON::Validator.validate(schema,data, :insert_defaults => true))
+    assert_valid schema,data, :insert_defaults => true
     assert_equal(42, data["a"])
     assert_equal(2, data[:b])
 
@@ -407,7 +408,7 @@ class Draft3Test < Minitest::Test
     data = {:b => 2}
     refute_valid schema, data
     assert_nil(data["a"])
-    assert(JSON::Validator.validate(schema,data, :insert_defaults => true))
+    assert_valid schema,data, :insert_defaults => true
     assert_equal(42, data["a"])
     assert_equal(2, data[:b])
 
@@ -423,7 +424,7 @@ class Draft3Test < Minitest::Test
     data = {:b => 2}
     refute_valid schema, data
     assert_nil(data["a"])
-    assert(!JSON::Validator.validate(schema,data, :insert_defaults => true))
+    refute_valid schema,data, :insert_defaults => true
     assert_nil(data["a"])
     assert_equal(2, data[:b])
 
@@ -439,7 +440,7 @@ class Draft3Test < Minitest::Test
     data = {:b => 2}
     assert_valid schema, data
     assert_nil(data["a"])
-    assert(!JSON::Validator.validate(schema,data, :insert_defaults => true))
+    refute_valid schema,data, :insert_defaults => true
     assert_equal("42",data["a"])
     assert_equal(2, data[:b])