+++ CHANGELOG.md
@@ -12,6 +12,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ### Fixed
 - Corrected the draft6 schema id to `http://json-schema.org/draft/schema#`
 - Rescue URI error when initializing a data string that contains a colon 
+- Made strict validation report errors for draft1 and draft2 schemas
 
 ## [2.8.0] - 2017-02-07